Refactor hacker dashboard with status as home page (#732)
* Integrated Sidebar into HackerDashboard

* Removed Team and Travel options from Sidebar temporarily

* Refactored HackerDashboard

* Commit old hacker dashboard file

* Rewrite return statement
