+++ src/features/Dashboard/Main.tsx
@@ -3,10 +3,9 @@ import { Redirect } from 'react-router';
 import { UserType } from '../../config';
 import { H1 } from '../../shared/Elements';
 import { getUserInfo } from '../../util';
-import StatusPage from '../Status/StatusPage';
-// import HackerDashboardContainer from './HackerDashboard';
 import SponsorDashboard from './SponsorDashboard';
 import StaffDashboardContainer from './StaffDashboard';
+import HackerDashboard from './HackerDashboard';
 
 interface IDashboardState {
   accountType: UserType;
@@ -40,7 +39,7 @@ class Dashboard extends React.Component<{}, IDashboardState> {
   public render() {
     switch (this.state.accountType) {
       case UserType.HACKER:
-        return <StatusPage />;
+        return <HackerDashboard />;
       case UserType.STAFF:
         return <StaffDashboardContainer />;
       case UserType.SPONSOR_T1:
